requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
google-generativeai>=0.5.0
google-cloud-firestore>=2.16.0
google-cloud-bigquery>=3.20.0
google-cloud-logging>=3.10.0
googlemaps>=4.10.0
firebase-admin>=6.5.0
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
